requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.24.0
pypdf>=4.0.0
Pillow>=10.0.0
pdfplumber>=0.11.0
pdf2docx>=0.5.8
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
reportlab>=4.1.0
python-barcode>=0.15.1
qrcode>=7.4.2
markdown>=3.5
psutil>=5.9.0
pyinstaller>=6.0.0
